Refactor Typescript Emitter and Emit TypeScriptClient
