+++ examples/npm-typescript/src/client.test.ts
@@ -1,0 +1,26 @@
+import { wirespecFetch, HandleFetch } from "wirespec/fetch";
+import { wirespecSerialization } from "wirespec/serialization";
+import {expect, test, vi} from "vitest";
+import {Client} from "./gen/client";
+
+// @ts-ignore
+const mockHandler = vi.fn<HandleFetch>(() => Promise.resolve({
+    status: 200,
+    headers: new Map([
+        ["Content-Type", "application/json"],
+        ["Content-Length", "2"],
+    ]),
+    text: () => Promise.resolve(`{"id": 123, "title": "test", "completed": false}`)
+}))
+
+test('testGetTodoById', async () => {
+    const client = Client(wirespecSerialization, (req) => wirespecFetch(req, mockHandler))
+    const res = await client.GetTodoById({id: "123"})
+    expect(res.status).toEqual(200)
+    expect(res.headers).toEqual({})
+    expect(res.body).toEqual({
+        "completed": false,
+        "id": 123,
+        "title": "test",
+    })
+})
