+++ examples/npm-typescript/src/clientProxy.test.ts
@@ -1,6 +1,7 @@
 import {GetTodoById, GetTodos, GetUsers, PostTodo} from "./gen/endpoint";
 import {Wirespec} from "./gen/Wirespec";
 import * as assert from "node:assert";
+import {expect, test} from "vitest";
 
 const serialization: Wirespec.Serialization = {
     deserialize<T>(raw: string | undefined): T {
@@ -66,21 +67,21 @@ const webClient: WebClient = (...apis) => {
 
 const api = webClient(PostTodo.api, GetTodos.api, GetTodoById.api, GetUsers.api)
 
-const testGetTodos = async () => {
+test('testGetTodos', async () => {
     const request: GetTodos.Request = GetTodos.request({});
     const response = await api.getTodos(request);
     const expected = {status: 200, headers: {total: 2}, body};
-    assert.deepEqual(response, expected);
-};
+    expect(response).toEqual(expected)
+});
 
-const testGetTodoById = async () => {
+test('testGetTodoById', async () => {
     const request: GetTodoById.Request = GetTodoById.request({id: "1"});
     const response = await api.getTodoById(request);
     const expected = GetTodoById.response200({body: body[0]});
-    assert.deepEqual(response, expected);
-};
+    expect(response).toEqual(expected)
+});
 
-const testPostTodo = async () => {
+test('testPostTodo', async () => {
     const request: PostTodo.Request = {
         method: "POST",
         path: {},
@@ -90,11 +91,5 @@ const testPostTodo = async () => {
     };
     const response = await api.postTodo(request);
     const expected = PostTodo.response200({body: {id: "3", name: "Do more", done: true}});
-    assert.deepEqual(response, expected);
-};
-
-Promise.all([
-    testGetTodos(),
-    testGetTodoById(),
-    testPostTodo()
-]);
+    expect(response).toEqual(expected)
+})
