+++ examples/npm-typescript/wirespec/user.ws
@@ -2,12 +2,13 @@ type User {
     name: String
 }
 
-type Error {
+type UserError {
     code: Integer,
     description: String
 }
 
 endpoint GetUsers GET /api/users -> {
     200 -> User[]
+    500 -> UserError
 }
