+++ src/compiler/core/src/commonMain/kotlin/community/flock/wirespec/compiler/core/emit/BaseEmitter.kt
@@ -1,0 +1,8 @@
+package community.flock.wirespec.compiler.core.emit
+
+interface BaseEmitter {
+
+    val extension: FileExtension
+
+    val shared: Shared?
+}
