+++ src/compiler/emitters/typescript/src/commonMain/kotlin/community/flock/wirespec/emitters/typescript/TypeScriptIdentifierEmitter.kt
@@ -1,0 +1,8 @@
+package community.flock.wirespec.emitters.typescript
+
+import community.flock.wirespec.compiler.core.emit.IdentifierEmitter
+import community.flock.wirespec.compiler.core.parse.Identifier
+
+interface TypeScriptIdentifierEmitter: IdentifierEmitter {
+    override fun emit(identifier: Identifier) = """"${identifier.value}""""
+}
