+++ src/plugin/npm/src/jsMain/resources/wirespec-serialization.d.ts
@@ -1,0 +1,3 @@
+export type Serialization = { serialize: <T>(typed: T) => string; deserialize: <T>(raw: string | undefined) => T }
+
+export const wirespecSerialization: Serialization
